Fixed error where MD5 was not actually being verified on downloaded files.  Updated from a minimum of CMake 2.8 to a much more specific 2.8.7.  Fixed CImg incorrect MD5.  Fixed error where system-wide CMake would be called instead of the active CMake install.
Turns out that the EXPECTED_MD5 subcommand to file() is only for fetching an MD5 from the server, if supported by said server, and comparing against the local file to see if the file actual needs to be downloaded. In my case that test is handled already another way.  However, to actually accomplish the MD5 test I needed a separate MD5 testing operation - turns out that has arrived as of 2.8.7.  Hence the much more specific version requirement.

Once in place, I found that CImg had an invalid MD5, so that got an update.  But before I could do so I had so update my own local CMake install - which I did in an unusual way that wound up exposing the fact that I was just blindly calling the system-wide cmake command instead of the one that was currently in use.  This caused a version incompatibility that was causing library extraction to not work.  Took way too long to hunt that one down due to the error reporting being a set instead of a message and therefore not actually reporting anything!