Fixed some more redudant includes.
Modularized EngineCore. I have moved most of main.cpp into EngineCore as logical
methods. This class will be turned into a static lib project next commit that allows
linking with native OS projects (such as Android, Win32, *nix, etc).

*NOTE* EngineCore::StartUp loads and runs config.as followed by a main.as. This
was to test 2 stage loading where in config.as all modules get loaded and registered
and then in main.as the various modules can be started as the start methods are now
registered. The name of this file will be changed soon to a non-hardcoded value.