Fixed error where Boost's auto-linker was attempting to find the incorrect libraries.
Since we provide our own Boost libs that are already ready to go, and already fed to the linker command, there's no need for Boost to attempt to auto-link - it will only get it wrong.

Also removed the compiler marker from the GetBoost.cmake Windows globs to match the actual file names.