Cleaned up Windows compile for the removed Boost Chrono.
Removing Boost Chrono's cimpiled side had the expected side-effect of forcing Chrono to be header-only.  Under the current state of the Mac edition of the engine that issue was hidden, but the Windows compilation choked.  Simply added the definitions needed to let Chrono know it needed to be header-only.  We don't use the hybrid error checking mode anyway!

Also had to remove the linker secs for Chrono and Thread from the Windows-specific CMakeLists.txt - for obvious reasons.